LA-91 Rename file extensions to indicate no longer using Extend Script and to avoid confusing reader/IDE/tools with React scripts. What comes closest is modern JS, hence using corresponding extension.
