+++ ArticleShapeExtractor/bootstrap.js
@@ -1,6 +1,6 @@
 require("./extensions/String.js");
 require("./extensions/globals.js");
-const Container = require("./modules/Container.inc.jsx");
+const Container = require("./modules/Container.js");
 
 function loadLocalConfigFile() {
     const fs = require("fs");
@@ -15,14 +15,14 @@ function loadLocalConfigFile() {
 }
 
 Container.registerSingleton("Settings", function() {
-    const Settings = require("./modules/Settings.inc.jsx");
-    require("./config/config.jsx");
+    const Settings = require("./modules/Settings.js");
+    require("./config/config.js");
     const localConfig = loadLocalConfigFile();
     return new Settings(globalThis.plaDefaultConfig, localConfig);
 });
 
 Container.registerSingleton("Logger", function() {
-    const Logger = require("./modules/Logger.inc.jsx");
+    const Logger = require("./modules/Logger.js");
     const config = Container.resolve("Settings").getLoggerConfig();
     var logger = new Logger(config.folder, config.filename, config.level, config.wipe);
     logger.init();
@@ -30,7 +30,7 @@ Container.registerSingleton("Logger", function() {
 });
 
 Container.registerSingleton("ErrorFactory", function() {
-    const ErrorFactory = require("./modules/ErrorFactory.inc.jsx");
+    const ErrorFactory = require("./modules/ErrorFactory.js");
     return new ErrorFactory(
         Container.resolve("Logger"),
         Container.resolve("Settings").getIncludeErrorDetailInAlerts(),
@@ -38,24 +38,24 @@ Container.registerSingleton("ErrorFactory", function() {
 });
 // Late initialize custom errors (once after ErrorFactory is registered).
 function delayedRequireErrorsModule() {
-    require("./modules/Errors.inc.jsx");
+    require("./modules/Errors.js");
 }
 delayedRequireErrorsModule();
 
 Container.registerSingleton("ArticleShapeGateway", function() {
-    const ArticleShapeGateway = require("./modules/ArticleShapeGateway.inc.jsx");
+    const ArticleShapeGateway = require("./modules/ArticleShapeGateway.js");
     return new ArticleShapeGateway(
         Container.resolve("Settings").getPlaServiceUrl(),
     );
 });
 
 Container.registerFactory("InDesignArticleService", function() {
-    const InDesignArticleService = require("./modules/InDesignArticleService.inc.jsx");
+    const InDesignArticleService = require("./modules/InDesignArticleService.js");
     return new InDesignArticleService();
 });
 
 Container.registerFactory("ExportInDesignArticlesToPlaService", function() {
-    const ExportInDesignArticlesToPlaService = require("./modules/ExportInDesignArticlesToPlaService.inc.jsx");
+    const ExportInDesignArticlesToPlaService = require("./modules/ExportInDesignArticlesToPlaService.js");
     const settings = Container.resolve("Settings");
     return new ExportInDesignArticlesToPlaService(
         Container.resolve("Logger"), 
@@ -67,7 +67,7 @@ Container.registerFactory("ExportInDesignArticlesToPlaService", function() {
 });
 
 Container.registerFactory("RegenerateArticleShapesService", function() {
-    const RegenerateArticleShapesService = require("./modules/RegenerateArticleShapesService.inc.jsx");
+    const RegenerateArticleShapesService = require("./modules/RegenerateArticleShapesService.js");
     return new RegenerateArticleShapesService(
         Container.resolve("Settings").getRegenerateArticleShapesQueryName(),
         Container.resolve("ExportInDesignArticlesToPlaService"),
