+++ ArticleShapeExtractor/commands/CreateLeadArticle.js
@@ -1,8 +1,8 @@
 CreateLeadArticle = function() {
     this.run = function() {
-        require('../bootstrap.jsx');
+        require('../bootstrap.js');
         try {
-            const Container = require("../modules/Container.inc.jsx");
+            const Container = require("../modules/Container.js");
             Container.resolve("InDesignArticleService").addOrRenameInDesignArticle("Lead");
         } catch (error) {
             error.alert();
