+++ ArticleShapeExtractor/commands/CreateSecondaryArticle.js
@@ -1,8 +1,8 @@
 CreateSecondaryArticle = function() {
     this.run = function() {
-        require('../bootstrap.jsx');
+        require('../bootstrap.js');
         try {
-            const Container = require("../modules/Container.inc.jsx");
+            const Container = require("../modules/Container.js");
             Container.resolve("InDesignArticleService").addOrRenameInDesignArticle("Secondary");
         } catch (error) {
             error.alert();
