+++ ArticleShapeExtractor/commands/CreateThirdArticle.js
@@ -1,8 +1,8 @@
 CreateThirdArticle = function() {
     this.run = function() {
-        require('../bootstrap.jsx');
+        require('../bootstrap.js');
         try {
-            const Container = require("../modules/Container.inc.jsx");
+            const Container = require("../modules/Container.js");
             Container.resolve("InDesignArticleService").addOrRenameInDesignArticle("Third");
         } catch (error) {
             error.alert();
