+++ ArticleShapeExtractor/config/config.js
