+++ ArticleShapeExtractor/modules/ArticleShapeGateway.js
