+++ ArticleShapeExtractor/modules/ErrorFactory.js
