+++ ArticleShapeExtractor/modules/ExportInDesignArticlesToPlaService.js
