+++ ArticleShapeExtractor/modules/InDesignArticleService.js
