+++ ArticleShapeExtractor/modules/Logger.js
