+++ ArticleShapeExtractor/modules/RegenerateArticleShapesService.js
