+++ ArticleShapeExtractor/modules/Settings.js
